整理代码

+++ lealone-cluster/src/main/java/org/lealone/cluster/utils/MurmurHash.java
@@ -17,8 +17,6 @@
  */
 package org.lealone.cluster.utils;
 
-import net.nicoulaj.compilecommand.annotations.Inline;
-
 import java.nio.ByteBuffer;
 
 /**
@@ -64,13 +62,13 @@ public class MurmurHash {
 
         if (left != 0) {
             if (left >= 3) {
-                h ^= (int) data.get(offset + length - 3) << 16;
+                h ^= data.get(offset + length - 3) << 16;
             }
             if (left >= 2) {
-                h ^= (int) data.get(offset + length - 2) << 8;
+                h ^= data.get(offset + length - 2) << 8;
             }
             if (left >= 1) {
-                h ^= (int) data.get(offset + length - 1);
+                h ^= data.get(offset + length - 1);
             }
 
             h *= m;
@@ -125,7 +123,7 @@ public class MurmurHash {
         case 2:
             h64 ^= (long) key.get(offset + length - rem + 1) << 8;
         case 1:
-            h64 ^= (long) key.get(offset + length - rem);
+            h64 ^= key.get(offset + length - rem);
             h64 *= m64;
         }
 
@@ -238,7 +236,7 @@ public class MurmurHash {
         case 2:
             k1 ^= ((long) key.get(offset + 1)) << 8;
         case 1:
-            k1 ^= ((long) key.get(offset));
+            k1 ^= (key.get(offset));
             k1 *= c1;
             k1 = rotl64(k1, 31);
             k1 *= c2;
